- Removed CONSTRUCTOR_ATTRIBUTE from __init_cpu_features declaration.
- Removed __get_aarch64_features.

- Split out changes for __arm_get_current_vg into a new file (sme-abi-vg.c)
  and removed from sme-abi-init.c.
